+++ RaccoonBitsCore/MathUtils.cs
@@ -2,7 +2,7 @@
 {
     public class MathUtils
     {
-        public static int Normalize(int value, int minInput, int maxInput, int minOutput, int maxOutput)
+        public static double Normalize(int value, int minInput, int maxInput, int minOutput, int maxOutput)
         {
             // Ensure the value is within the input range
             value = Math.Max(minInput, Math.Min(maxInput, value));
@@ -10,7 +10,7 @@
             // Perform linear normalization
             double normalizedValue = (value - minInput) / (double)(maxInput - minInput) * (maxOutput - minOutput) + minOutput;
 
-            return (int)normalizedValue;
+            return normalizedValue;
         }
     }
 }
